% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-ED-AP.R
\name{ap_from_ed_sent}
\alias{ap_from_ed_sent}
\title{Admissions from Emergency Departments, with evidence in presentation record}
\usage{
ap_from_ed_sent(
  arrival,
  actual_departure,
  modesep_presentation,
  facility_emergency,
  episode_start,
  facility_admission,
  facilities_must_match = TRUE,
  admission_codes = !!getOption("jnyr.eddc.modesep.admission"),
  admission_time = !!getOption("jnyr.eddc.time.admission"),
  pre_presentation_buffer = !!getOption("jnyr.eddc.time.buffer")
)
}
\arguments{
\item{arrival}{The emergency presentation arrival timestamp.}

\item{actual_departure}{The actual departure timestamp from the emergency
department.}

\item{modesep_presentation}{The mode of separation for the end of the
emergency presentation.}

\item{facility_emergency}{The recorded facility in the emergency presentation
record.}

\item{episode_start}{The timestamp of the admitted patient episode start.}

\item{facility_admission}{The recorded facility of the admitted patient
record.}

\item{facilities_must_match}{If true, then \code{facility_admission} must match
\code{facility_emergency} for a link to be considered.}

\item{admission_codes}{The codes in \code{modesep_presentation} that correspond to
an admission.}

\item{admission_time}{Time allowed from the end of presentation to the
admission starting.}

\item{pre_presentation_buffer}{Fudge factor: how much earlier can the
admission start before the presentation arrival for a link to be
considered. Only included to counter data quality issues.}
}
\value{
An \code{rlang::expr} for use in \code{find_links}.
}
\description{
\code{ap_from_ed_sent} finds transfers where a patient was admitted from an
emergency department presentation. The evidence for these transfers is taken
from the mode of separation at the end of the presentation.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
