% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_pohcs.R
\name{collect_journeys}
\alias{collect_journeys}
\title{Collect patient journeys from links}
\usage{
collect_journeys(
  data,
  links,
  identifier,
  links_from = character(),
  links_to = character(),
  journey = "journey"
)
}
\arguments{
\item{data}{The data to link together, optionally as a list of \code{data.table}s.}

\item{links}{A two-column table of links, or a list of such tables if \code{data}
is a list. The first column should indicate parent episodes, and the second
should indicate child episodes.}

\item{identifier}{The name of the column with episode identifiers.}

\item{links_from}{(If \code{data} is a list.) Which item of \code{data} corresponds to
the parent columns in each item in \code{links}.}

\item{links_to}{(If \code{data} is a list.) Which item of \code{data} corresponds to
the child columns in each item in \code{links}.}

\item{journey}{The name of the journey column to create.}
}
\value{
\code{collect_journeys} adds the column \code{journey} to the original data for
the journey sequence number. It then returns the updated data, invisibly.
}
\description{
\code{collect_journeys} takes a dataset and the links between records, and adds a
column for the patient journey number. This number is unique for every
journey, not just journeys within the same person.
}
\details{
For longer-running collections, \code{collect_journeys} will display a progress
bar. If this does not appear, then set the RSTUDIO environment variable to
'1' with \code{Sys.setenv(RSTUDIO = '1')}. If you want to disable this, set the
\code{progress_enabled} option to FALSE with \code{options(progress_enabled = FALSE)}.
}
