% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_pred_options}
\alias{default_pred_options}
\title{Default options for predicates by jurisdiction}
\format{
A list with two levels. The first level defines a jurisdiction, and
the second level defines the options for the package. Currently, it
contains options defined for:
\itemize{
\item NSW
}
}
\usage{
default_pred_options
}
\description{
\code{default_options} is a list that contains parameters for various predicates
by jurisdiction. These can be used to define default parameters in a flexible
manner.
}
\examples{
options(default_pred_options$NSW)
transfers_both(
  episode_end_parent, mode_of_separation_recode_parent,
  facility_identifier_recode_parent, facility_trans_to_recode_parent,
  episode_start_child,
  facility_identifier_recode_child, facility_trans_from_child
)
}
