% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-ED-AP.R
\name{ed_from_ap_sent}
\alias{ed_from_ap_sent}
\title{Transfers from an admitted patient episode to an emergency department}
\usage{
ed_from_ap_sent(
  arrival,
  episode_end,
  modesep_admission,
  transfer_time = !!getOption("jnyr.apdc.time.transfer"),
  transfer_codes = !!getOption("jnyr.apdc.modesep.edtrans")
)
}
\arguments{
\item{arrival}{The emergency presentation arrival timestamp.}

\item{episode_end}{The admission episode end timestamp.}

\item{modesep_admission}{The mode of separation from the admission record.}

\item{transfer_time}{The time allowed for a transfer to take place}

\item{transfer_codes}{Codes for \code{modesep_admission} that indicate a transfer
has taken place.}
}
\value{
An \code{rlang::expr} for use in \code{find_links}.
}
\description{
\code{ed_from_ap_sent} identifies presentations to an emergency department where
patients were transferred from an admitted patient episode.
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
