% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-ED-ED.R
\name{ed_transfers_sent}
\alias{ed_transfers_sent}
\title{Transfers sent by an emergency department.}
\usage{
ed_transfers_sent(
  actual_departure_parent,
  modesep_parent,
  arrival_child,
  transfer_time = !!getOption("jnyr.eddc.time.transfer"),
  transfer_codes = !!getOption("jnyr.eddc.modesep.transfer")
)
}
\arguments{
\item{actual_departure_parent}{The actual departure timestamp from the
candidate parent presentation.}

\item{modesep_parent}{The recorded mode of separation in the candidate parent
presentation.}

\item{arrival_child}{The arrival timestamp from the candidate child
presentation.}

\item{transfer_time}{Time allowed for transfers between emergency
departments.}

\item{transfer_codes}{The codes in \code{modesep_parent} that indicate an
inter-department transfer may have taken place.}
}
\value{
An \code{rlang::expr} for use in \code{find_links}.
}
\description{
\code{ed_transfers_sent} considers two presentations linked where the recorded
mode of separation for a presentation indicates departure for another
hospital, emergency department or health service, and there is a presentation
starting within six hours after departure. The maximum time allowed between
presentations is adjustable, as are the codes for transfer.
}
\references{
\itemize{
\item Mode of separation codes:
\href{http://hird.health.nsw.gov.au/hird/view_domain_values.cfm?ItemID=10803}{}
}
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
