% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{find_links}
\alias{find_links}
\title{Find links based on predicates using \code{dbplyr}}
\usage{
find_links(.data, ..., .keep = FALSE)
}
\arguments{
\item{.data}{The input data to filter, which should usually be a lazy \code{tbl}
from the \code{dbplyr} package.}

\item{...}{Expressions that return a logical value, in terms of the variables
of \code{.data}. Links are returned if they satisfy ANY of these expressions,
which differs from the behaviour of, for example, \code{dplyr::filter}.}

\item{.keep}{If this is true, then the columns of predicates are retained in
the output data. By default, they are dropped.}
}
\value{
\code{find_links} returns an object of the same type as \code{.data}.
}
\description{
\code{find_links} is used to find links between datasets according to a number of
user-specified predicates. By default, \code{dplyr::filter} takes the intersection
of all arguments, whereas this function returns the union.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
