% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journeyer-package.R
\docType{package}
\name{journeyer-package}
\alias{journeyer}
\alias{journeyer-package}
\title{journeyer: Patient Journeys through Administrative Data}
\description{
Collect patient journeys with customisable definitions using dbplyr.
}
\author{
\strong{Maintainer}: Michael Nelson \email{Michael.Nelson1@health.nsw.gov.au} [reviewer, contributor]

Authors:
\itemize{
  \item Ian Powell \email{Ian.Powell1@health.nsw.gov.au}
  \item Zhisheng Sa \email{Zhisheng.Sa@health.nsw.gov.au}
}

Other contributors:
\itemize{
  \item Branislav Igic \email{branislav.igic@health.nsw.gov.au} [contributor]
  \item Maria Alfaro-Ramirez \email{maria.alfaroramirez@health.nsw.gov.au} [contributor]
  \item State of New South Wales NSW Ministry of Health \email{MOH-CEE@health.nsw.gov.au} [copyright holder]
}

}
\keyword{internal}
