% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-ED-ED.R
\name{overlap_presentations}
\alias{overlap_presentations}
\title{Emergency presentations that overlap in time}
\usage{
overlap_presentations(arrival_parent, actual_departure_parent, arrival_child)
}
\arguments{
\item{arrival_parent, actual_departure_parent}{The start and end timestamps in
the candidate parent presentation.}

\item{arrival_child}{The start timestamp of the candidate child presentation.}
}
\value{
An \code{rlang::expr} for use in \code{find_links}.
}
\description{
\code{overlap_presentations} finds presentations that have some overlap time. It
checks to see if the child presentation starts during the parent presentation
(inclusive of start and end times).
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
