% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{sameday_transfer}
\alias{sameday_transfer}
\title{Explicit transfers within and between hospitals}
\usage{
sameday_transfer(
  episode_start_parent,
  episode_end_parent,
  modesep_parent,
  episode_start_child,
  transfer_codes = !!getOption("jnyr.apdc.modesep.sameday")
)
}
\arguments{
\item{episode_start_parent, episode_end_parent}{The start and end timestamps
in the candidate parent episode.}

\item{modesep_parent}{The recorded mode of separation in the
candidate parent episode.}

\item{episode_start_child}{The start timestamp of the candidate
child episode.}

\item{transfer_codes}{The codes in \code{modesep_parent} that correspond to
transfers in the same day.}
}
\value{
\code{sameday_transfer} returns a call to be used in \code{dplyr}/\code{dbplyr}
filters.
}
\description{
\code{sameday_transfer} identifies transfers where the source/parent episode
explicitly ends with a type-change or transfer to another hospital, and the
destination/child episode starts on the same day.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
