% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{transfers_in}
\alias{transfers_in}
\title{Transfers where child episode has evidence of transfer}
\usage{
transfers_in(
  episode_end_parent,
  facility_parent,
  episode_start_child,
  source_referral_child,
  facility_from_child,
  referral_codes = !!getOption("jnyr.apdc.refer.transfers"),
  transfer_time = !!getOption("jnyr.apdc.time.transfer"),
  facilities_must_match = TRUE
)
}
\arguments{
\item{episode_end_parent}{End timestamp of candidate parent episode.}

\item{facility_parent}{Facility of candidate parent episode.}

\item{episode_start_child}{Start timestamp of candidate child episode.}

\item{source_referral_child}{The source of referral for the child episode.}

\item{facility_from_child}{In the child episode, the recorded facility the
patient was transferred from.}

\item{referral_codes}{The source of referral codes that correspond to a
patient being transferred into an episode.}

\item{transfer_time}{The time allowed between episodes for a transfer to take
place, in a form suitable for your database management system.}

\item{facilities_must_match}{Does \code{facility_from_child} have to match
\code{facility_parent} for a transfer to be indicated?}
}
\value{
\code{transfers_in} returns a call to be used in \code{dplyr}/\code{dbplyr} filters.
}
\description{
\code{transfers_in} finds transfers between episodes, where the child episode has
evidence of a transfer coming in. In this function, the parent episode does
not require evidence of a transfer going out. However, by default the parent
episode's facility must match the child episode's facility transferred from,
but this can be disabled with the \code{facilities_must_match} argument.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
