% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{transfers_out}
\alias{transfers_out}
\title{Transfers where parent episode has evidence of transfer}
\usage{
transfers_out(
  episode_end_parent,
  modesep_parent,
  facility_to_parent,
  episode_start_child,
  facility_child,
  transfer_codes = !!getOption("jnyr.apdc.modesep.transfers"),
  transfer_time = !!getOption("jnyr.apdc.time.transfer"),
  facilities_must_match = TRUE
)
}
\arguments{
\item{episode_end_parent}{End timestamp of candidate parent episode.}

\item{modesep_parent}{The mode of separation of the candidate parent episode.}

\item{facility_to_parent}{The destination facility transferred to, as
recorded in the candidate parent episode.}

\item{episode_start_child}{Start timestamp of candidate child episode.}

\item{facility_child}{Facility of the candidate child episode.}

\item{transfer_codes}{Mode of separation codes corresponding to a transfer.}

\item{transfer_time}{The time allowed between episodes for a transfer to take
place, in a form suitable for your database management system.}

\item{facilities_must_match}{Does \code{facility_from_child} have to match
\code{facility_parent} for a transfer to be indicated?}
}
\value{
\code{transfers_in} returns a call to be used in \code{dplyr}/\code{dbplyr} filters.
}
\description{
\code{transfers_out} finds transfers between episodes, where the parent episode
has evidence of a transfer coming in. In this function, the child episode
does not require evidence of a transfer coming in. However, by default the
child episode's facility must match the parent episode's facility transferred
to, but this can be disabled with the \code{facilities_must_match} argument.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
