% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{type_change}
\alias{type_change}
\title{Type-change separations and admissions}
\usage{
type_change(
  episode_end_parent,
  modesep_parent,
  facility_parent,
  episode_start_child,
  source_referral_child,
  facility_child,
  separation_codes = !!getOption("jnyr.apdc.modesep.tcs"),
  referral_codes = !!getOption("jnyr.apdc.refer.tcs"),
  transfer_time = !!getOption("jnyr.apdc.time.tcs"),
  facilities_must_match = TRUE,
  require = c("both", "separation", "referral")
)
}
\arguments{
\item{episode_end_parent}{End timestamp of candidate parent episode.}

\item{modesep_parent}{The mode of separation of the candidate parent episode.}

\item{facility_parent}{Facility of candidate parent episode.}

\item{episode_start_child}{Start timestamp of candidate child episode.}

\item{source_referral_child}{The source of referral for the child episode.}

\item{facility_child}{Facility of the candidate child episode.}

\item{separation_codes}{Mode of separation codes corresponding to a type
change.}

\item{referral_codes}{The source of referral codes that correspond to a
patient starting an episode with a type change.}

\item{transfer_time}{The time allowed between episodes for a type change to
take place, in a form suitable for your database management system.}

\item{facilities_must_match}{Does \code{facility_parent} have to match
\code{facility_child} for a link to be eligible for a type change?}

\item{require}{Whether evidence of a either a type change admission,
separation, or both is required.}
}
\value{
\code{type_change} returns a call to be used in \code{dplyr}/\code{dbplyr} filters.
}
\description{
\code{type_change} finds type changes which correspond to a new episode in the
same period of hospital care where a patient's type of care has changed. This
change in care can result in a new record. By default, the parent episode
must end in a type change separation (TCS), and the child episode must start
with a type change admission (TCA), and both episodes must be in the same
facility.
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()}
}
\concept{link predicates}
