% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AMB.R
\name{amb_arrival}
\alias{amb_arrival}
\title{Ambulance arrivals at emergency departments or hospitalisations}
\usage{
amb_arrival(
  ambulance_start,
  ambulance_end,
  child_start,
  buffer_time = !!getOption("jnyr.amb.time.buffer")
)
}
\arguments{
\item{ambulance_start, ambulance_end}{Start and end timestamps of the
candidate parent ambulance call, respectively.}

\item{child_start}{Start timestamp of the candidate child presentation or
admission.}

\item{buffer_time}{Time allowed between the end of an ambulance call and the
start of a presentation/admission.}
}
\value{
\code{amb_arrival} returns a call to be used in \code{dplyr}/\code{dbplyr} filters.
}
\description{
\code{amb_arrival} finds links between ambulance calls and either emergency
department presentations or admitted patient episodes. By default, an hour is
allowed after the ambulance call for the presentation or episode to start.
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
