% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AMB.R
\name{amb_transfer_out}
\alias{amb_transfer_out}
\title{Transfers out of hospital using an ambulance}
\usage{
amb_transfer_out(
  ambulance_start,
  ambulance_end,
  episode_end,
  modesep,
  buffer_time = !!getOption("jnyr.amb.time.buffer"),
  transfer_codes = !!getOption("jnyr.apdc.modesep.transfers")
)
}
\arguments{
\item{ambulance_start, ambulance_end}{Start and end timestamps of the
candidate ambulance call, respectively.}

\item{episode_end}{End timestamp of the candidate hospitalisation
transferring the patient out.}

\item{modesep}{Recorded mode of separation of the candidate admission.}

\item{buffer_time}{Time allowed before the end of an admission that an
ambulance call qualifies.}

\item{transfer_codes}{Mode of separation codes corresponding to a transfer.}
}
\value{
\code{amb_transfer_out} returns a call to be used in \code{dplyr}/\code{dbplyr}
filters.
}
\description{
\code{amb_transfer_out} finds ambulance calls which follow the end of an admitted
patient episode, where the patient is recorded as being transferred to
another location. By default, an ambulance call can start up to an hour
before the end of an admission.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
