% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R, R/zzz.R
\name{default_predicates}
\alias{default_predicates}
\title{Default predicates for period of care links by jurisdiction}
\format{
A \code{list} with three levels:
\enumerate{
\item The first level defines a jurisdiction.
\item The second level defines the datasets which the predicates find links
between.
\item The third level contains the predicates as expressions.
}

In each case, the predicates assume that the datasets are joined using an
inner join, and where there are column name collisions, the first dataset
has suffix \verb{_parent}, and the second has suffix \verb{_child}. They also assume
that the correct options have been set, for example with
\code{default_pred_options}.
}
\usage{
default_predicates
}
\description{
\code{default_predicates} is a list that contains default predicates
for a variety of datasets within jurisdictions.
}
\section{NSW}{

\subsection{APDC-APDC links}{

The predicates to find links within the NSW APDC are:
\itemize{
\item \code{\link{overlap_episodes}}
\item \code{\link{transfers_in}}, with facilities required to match
\item \code{\link{transfers_out}}, with facilities required to match
\item \code{\link{transfers_both}}, where facilities do not have to match
\item \code{\link{type_change}}, where parent ends with a type change
separation and child starts with type change admission in the same
facility
}
}

\subsection{EDDC-APDC links}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The predicates to find links between the NSW EDDC and the NSW APDC are:
\itemize{
\item \code{\link{overlap_episodes}}, where the admission starts during a
presentation
\item \code{\link{overlap_episodes}}, where the presentation starts during
an admission
\item \code{\link{ap_from_ed_received}}, where presentation and admission
must be in the same hospital
\item \code{\link{ap_from_ed_sent}}, where the presentation and admission
must be in the same hospital
\item \code{\link{ap_from_ed_both}}, where the presentation and admission
can be in separate hospitals
\item \code{\link{ed_from_ap_sent}}
}
}

\subsection{EDDC-EDDC links}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The predicates to find links within the NSW EDDC are:
\itemize{
\item \code{\link{ed_transfers_sent}}
\item \code{\link{ed_transfers_received}}
\item \code{\link{overlap_presentations}}
}
}
}

