% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-ED-ED.R
\name{ed_transfers_received}
\alias{ed_transfers_received}
\title{Transfers received by an emergency department.}
\usage{
ed_transfers_received(
  actual_departure_parent,
  arrival_child,
  source_of_referral_child,
  transfer_time = !!getOption("jnyr.eddc.time.transfer"),
  referral_codes = !!getOption("jnyr.eddc.refer.transfer")
)
}
\arguments{
\item{actual_departure_parent}{The actual departure timestamp from the
candidate parent presentation.}

\item{arrival_child}{The arrival timestamp from the candidate child
presentation.}

\item{source_of_referral_child}{The recorded source of referral code in the
candidate child presentation.}

\item{transfer_time}{Time allowed for transfers between emergency
departments.}

\item{referral_codes}{The codes in \code{source_of_referral_child} that indicate
an inter-department transfer may have taken place.}
}
\value{
An \code{rlang::expr} for use in \code{find_links}.
}
\description{
\code{ed_transfers_received} considers two presentations linked where the recorded
source of referral for a presentation is another hospital or an emergency
department, and there is a presentation ending soon earlier. The maximum time
allowed between presentations is adjustable, as are the codes for transfer.
}
\references{
\itemize{
\item Source of referral codes:
\href{http://hird.health.nsw.gov.au/hird/view_domain_values.cfm?ItemID=10810}{}
}
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=samestart]{samestart()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
