% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{LinkPredicates}
\alias{LinkPredicates}
\title{Episode linkage predicate calls}
\description{
Episode linkage predicate calls identify which episodes have transfers
between them. All of these calls assume that the person is the same for each
episode, meaning they can be included in a query after an inner join on PPN.
These episode linkage predicates are designed for use with \code{dplyr} and
\code{dbplyr} interfacing with SQL.
}
\section{Inputs}{


The inputs to any SQL episode call generator should be unquoted columns,
possibly operated on by database-internal SQL macros. See
\code{\link[dbplyr]{partial_eval}} in the \code{dbplyr} package for more
information.
}

\section{Value}{


Every episode linkage function should return a call which, when evaluated
by \code{dbplyr}, returns true if a pair of records with matching PPNs has a
transfer between them, and false otherwise.
}

\seealso{
Other link predicates: 
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
