% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{overlap_episodes}
\alias{overlap_episodes}
\title{Episodes that overlap in time}
\usage{
overlap_episodes(episode_start_parent, episode_end_parent, episode_start_child)
}
\arguments{
\item{episode_start_parent, episode_end_parent}{The start and end timestamps
in the candidate parent episode.}

\item{episode_start_child}{The start timestamp of the candidate
child episode.}
}
\value{
\code{overlap_episodes} returns a call to be used in \code{dplyr}/\code{dbplyr}
filters.
}
\description{
\code{overlap_episodes} finds episodes that have some overlap in their admission
times. It checks to see if the child episode starts during the parent episode
(inclusive of start and end times).
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_both}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
