% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{samestart}
\alias{samestart}
\title{Episodes with same start date and time}
\usage{
samestart(
  episode_start_parent,
  episode_end_parent,
  episode_start_child,
  episode_end_child
)
}
\arguments{
\item{episode_start_parent, episode_end_parent}{The start and end timestamps
in the candidate parent episode.}

\item{episode_start_child, episode_end_child}{The start and end timestamps in
the candidate child episode.}
}
\value{
\code{samestart} returns a call to be used in \code{dplyr}/\code{dbplyr} filters.
}
\description{
\code{samestart} identifies transfers where two episodes have the exact same start
date and time. The child episode finishes on or before the end of the parent
episode.
}
\seealso{
Other link predicates:
\code{\link{LinkPredicates}},
\code{\link[=amb_arrival]{amb_arrival()}},
\code{\link[=amb_transfer_out]{amb_transfer_out()}},
\code{\link[=ap_from_ed_both]{ap_from_ed_both()}},
\code{\link[=ap_from_ed_received]{ap_from_ed_received()}},
\code{\link[=ap_from_ed_sent]{ap_from_ed_sent()}},
\code{\link[=ed_from_ap_sent]{ed_from_ap_sent()}},
\code{\link[=ed_transfers_received]{ed_transfers_received()}},
\code{\link[=ed_transfers_sent]{ed_transfers_sent()}},
\code{\link[=find_links]{find_links()}},
\code{\link[=overlap_episodes]{overlap_episodes()}},
\code{\link[=overlap_presentations]{overlap_presentations()}},
\code{\link[=sameday_transfer]{sameday_transfer()}},
\code{\link[=transfers_both]{transfers_both()}},
\code{\link[=transfers_in]{transfers_in()}},
\code{\link[=transfers_out]{transfers_out()}},
\code{\link[=type_change]{type_change()}}
}
\concept{link predicates}
