% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-AP-AP.R
\name{transfers_both}
\alias{transfers_both}
\title{Transfers where both parent and child episodes have evidence of transfer}
\usage{
transfers_both(
  episode_end_parent,
  modesep_parent,
  facility_parent,
  facility_to_parent,
  episode_start_child,
  source_referral_child,
  facility_child,
  facility_from_child,
  referral_codes = !!getOption("jnyr.apdc.refer.transfers"),
  transfer_codes = !!getOption("jnyr.apdc.modesep.transfers"),
  transfer_time = !!getOption("jnyr.apdc.time.transfer"),
  facilities_must_match = FALSE
)
}
\arguments{
\item{episode_end_parent}{End timestamp of candidate parent episode.}

\item{modesep_parent}{The mode of separation of the candidate parent episode.}

\item{facility_parent}{Facility of candidate parent episode.}

\item{facility_to_parent}{The destination facility transferred to, as
recorded in the candidate parent episode.}

\item{episode_start_child}{Start timestamp of candidate child episode.}

\item{source_referral_child}{The source of referral for the child episode.}

\item{facility_child}{Facility of the candidate child episode.}

\item{facility_from_child}{In the child episode, the recorded facility the
patient was transferred from.}

\item{referral_codes}{The source of referral codes that correspond to a
patient being transferred into an episode.}

\item{transfer_codes}{Mode of separation codes corresponding to a transfer.}

\item{transfer_time}{The time allowed between episodes for a transfer to take
place, in a form suitable for your database management system.}

\item{facilities_must_match}{Does \code{facility_from_child} have to match
\code{facility_parent} for a transfer to be indicated?}
}
\value{
\code{transfers_in} returns a call to be used in \code{dplyr}/\code{dbplyr} filters.
}
\description{
\code{transfers_both} finds transfers between episodes, where the parent episode
has evidence of a transfer coming in and the child episode has evidence of a
transfer going out. Optionally, this function can require the appropriate
facility columns to match, but by default does not.
}
\seealso{
Other link predicates: 
\code{\link{LinkPredicates}},
\code{\link{amb_arrival}()},
\code{\link{amb_transfer_out}()},
\code{\link{ap_from_ed_both}()},
\code{\link{ap_from_ed_received}()},
\code{\link{ap_from_ed_sent}()},
\code{\link{ed_from_ap_sent}()},
\code{\link{ed_transfers_received}()},
\code{\link{ed_transfers_sent}()},
\code{\link{find_links}()},
\code{\link{overlap_episodes}()},
\code{\link{overlap_presentations}()},
\code{\link{sameday_transfer}()},
\code{\link{samestart}()},
\code{\link{transfers_in}()},
\code{\link{transfers_out}()},
\code{\link{type_change}()}
}
\concept{link predicates}
